More precise types

+++ examples/suitup/manifest-parser.h
@@ -1,9 +1,9 @@
 typedef struct manifest_s {
-    int versionID;
-    int sequenceNumber;
+    uint8_t versionID;
+    uint32_t sequenceNumber;
     struct condition_s *preConditions;
     struct condition_s *postConditions;
-    int contentKeyMethod;
+    uint8_t contentKeyMethod;
     struct payloadInfo_s *payloadInfo;
     struct URLDigest_s *precursorImage;
     struct URLDigest_s *dependencies;
@@ -11,15 +11,15 @@ typedef struct manifest_s {
 } manifest_t;
 
 typedef struct condition_s {
-    int type;
+    int8_t type;
     char *value;
     struct condition_s *next;
 } condition_t;
 
 typedef struct payloadInfo_s {
-    int format;
-    int size;
-    int storage;
+    uint8_t format;
+    uint32_t size;
+    uint8_t storage;
     struct URLDigest_s *URLDigest;
 } payloadInfo_t;
 
@@ -30,13 +30,13 @@ typedef struct URLDigest_s {
 } URLDigest_t;
 
 typedef struct option_s {
-    int type;
+    int8_t type;
     char *value;
     struct option_s *next;
 } option_t;
 
 void manifest_parser(manifest_t*, char*);
-int get_next_key(char**);
+uint8_t get_next_key(char**);
 char *get_next_value(char**);
-int is_digit(char*);
+uint8_t is_digit(char*);
 void print_manifest(manifest_t*);
